+++ pkg/ffuf/optionsparser.go
@@ -37,6 +37,7 @@ type HTTPOptions struct {
 	RecursionDepth    int
 	RecursionStrategy string
 	ReplayProxyURL    string
+	SNI               string
 	Timeout           int
 	URL               string
 }
@@ -49,6 +50,7 @@ type GeneralOptions struct {
 	Delay                  string
 	MaxTime                int
 	MaxTimeJob             int
+	Noninteractive         bool
 	Quiet                  bool
 	Rate                   int
 	ShowVersion            bool `toml:"-"`
@@ -73,11 +75,11 @@ type InputOptions struct {
 }
 
 type OutputOptions struct {
-	DebugLog              string
-	OutputDirectory       string
-	OutputFile            string
-	OutputFormat          string
-	OutputCreateEmptyFile bool
+	DebugLog            string
+	OutputDirectory     string
+	OutputFile          string
+	OutputFormat        string
+	OutputSkipEmptyFile bool
 }
 
 type FilterOptions struct {
@@ -112,6 +114,7 @@ func NewConfigOptions() *ConfigOptions {
 	c.General.Delay = ""
 	c.General.MaxTime = 0
 	c.General.MaxTimeJob = 0
+	c.General.Noninteractive = false
 	c.General.Quiet = false
 	c.General.Rate = 0
 	c.General.ShowVersion = false
@@ -130,6 +133,7 @@ func NewConfigOptions() *ConfigOptions {
 	c.HTTP.RecursionStrategy = "default"
 	c.HTTP.ReplayProxyURL = ""
 	c.HTTP.Timeout = 10
+	c.HTTP.SNI = ""
 	c.HTTP.URL = ""
 	c.Input.DirSearchCompat = false
 	c.Input.Extensions = ""
@@ -148,7 +152,7 @@ func NewConfigOptions() *ConfigOptions {
 	c.Output.OutputDirectory = ""
 	c.Output.OutputFile = ""
 	c.Output.OutputFormat = "json"
-	c.Output.OutputCreateEmptyFile = false
+	c.Output.OutputSkipEmptyFile = false
 	return c
 }
 
@@ -251,6 +255,11 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 		conf.Url = parseOpts.HTTP.URL
 	}
 
+	// Prepare SNI
+	if parseOpts.HTTP.SNI != "" {
+		conf.SNI = parseOpts.HTTP.SNI
+	}
+
 	//Prepare headers and make canonical
 	for _, v := range parseOpts.HTTP.Headers {
 		hs := strings.SplitN(v, ":", 2)
@@ -259,14 +268,14 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 			// except if used in custom defined header
 			var CanonicalNeeded = true
 			for _, a := range conf.CommandKeywords {
-				if a == hs[0] {
+				if strings.Contains(hs[0], a) {
 					CanonicalNeeded = false
 				}
 			}
 			// check if part of InputProviders
 			if CanonicalNeeded {
 				for _, b := range conf.InputProviders {
-					if b.Keyword == hs[0] {
+					if strings.Contains(hs[0], b.Keyword) {
 						CanonicalNeeded = false
 					}
 				}
@@ -384,7 +393,7 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 	conf.InputShell = parseOpts.Input.InputShell
 	conf.OutputFile = parseOpts.Output.OutputFile
 	conf.OutputDirectory = parseOpts.Output.OutputDirectory
-	conf.OutputCreateEmptyFile = parseOpts.Output.OutputCreateEmptyFile
+	conf.OutputSkipEmptyFile = parseOpts.Output.OutputSkipEmptyFile
 	conf.IgnoreBody = parseOpts.HTTP.IgnoreBody
 	conf.Quiet = parseOpts.General.Quiet
 	conf.StopOn403 = parseOpts.General.StopOn403
@@ -399,6 +408,7 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 	conf.Timeout = parseOpts.HTTP.Timeout
 	conf.MaxTime = parseOpts.General.MaxTime
 	conf.MaxTimeJob = parseOpts.General.MaxTimeJob
+	conf.Noninteractive = parseOpts.General.Noninteractive
 	conf.Verbose = parseOpts.General.Verbose
 
 	// Handle copy as curl situation where POST method is implied by --data flag. If method is set to anything but GET, NOOP
@@ -490,6 +500,12 @@ func parseRawRequest(parseOpts *ConfigOptions, conf *Config) error {
 	}
 	conf.Data = string(b)
 
+	// Remove newline (typically added by the editor) at the end of the file
+	if strings.HasSuffix(conf.Data, "\r\n") {
+		conf.Data = conf.Data[:len(conf.Data)-2]
+	} else if strings.HasSuffix(conf.Data, "\n") {
+		conf.Data = conf.Data[:len(conf.Data)-1]
+	}
 	return nil
 }