+++ pkg/output/file_json.go
@@ -38,11 +38,6 @@ type jsonFileOutput struct {
 }
 
 func writeEJSON(filename string, config *ffuf.Config, res []ffuf.Result) error {
-
-	if config.OutputCreateEmptyFile && (len(res) == 0) {
-		return nil
-	}
-
 	t := time.Now()
 	outJSON := ejsonFileOutput{
 		CommandLine: config.CommandLine,